add-stationary-freq
Added the attribute self.stat_freq to store the marginal frequency of each alph in the alphabet, added the attribute self.stat_prob to store the stationary proabilities of each alph in the alphabet, and added the method self.freq3() to the GenMarkovTransitionProb class to return the frequency of alph.

Added script to tally total frequencies of each alph within the __init__ function.